# PERF: faster access to the dtype for masked numeric arrays